requirements.txt
Django==3.2.5
django-crispy-forms
django-extensions
openpyxl
dateparser
